updated main.ipynb to include final questions
Something is still going wrong with the final two exercises. But before fixing I want to speed up the code. Apparently a full training cycle can be done in 2 min.

As a start I created another replay buffer that assigns a np.array rather than a list with pop and append